refactor(teams): split triggers module [OPE-263]
